Added Profile and Choices fragment and xml
