+++ app/src/main/res/drawable/capybara.jpg
