+++ app/src/main/res/layout/fragment_profile.xml
@@ -3,12 +3,40 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileFragment">
+    tools:context=".ProfileFragment"
+    android:background="#FFC0CB">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="Profile Fragment" />
+        android:id="@+id/topLeftText"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="top|start"
+        android:layout_marginTop="25dp"
+        android:text="-User Profile-"
+        android:textColor="@color/black"
+        android:textSize="24dp"
+        android:textStyle="bold"/>
 
+    <ImageView
+        android:id="@+id/pfpimage"
+        android:layout_width="150dp"
+        android:layout_height="100dp"
+        android:layout_gravity="center_horizontal"
+        android:layout_marginTop="80dp"
+        android:src="@drawable/capybara"
+        android:contentDescription="Capybara Profile Image"
+    />
+
+    <TextView
+        android:id="@+id/nameText"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="20dp"
+        android:textColor="@color/black"
+        android:layout_gravity="center_horizontal"
+        android:text="A Capybara"
+        android:textStyle="bold"
+        android:layout_marginTop="180dp"
+    />
 </FrameLayout>
