--- fragment_choices.xml
@@ -1,93 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ChoicesFragment"
-    android:background="#FFC0CB">
-
-    <!-- TODO: Update blank fragment layout -->
-    <!-- Watched this to figure out the radio buttons -->
-    <!--https://www.youtube.com/watch?v=IdPeTllyebA&ab_channel=CodingSTUFF-->
-    <!-- This for spinner/dropdown https://developer.android.com/develop/ui/views/components/spinner-->
-
-    <TextView
-        android:id="@+id/topLeftText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top|start"
-        android:layout_marginTop="25dp"
-        android:text="-Choices Screen-"
-        android:textColor="@color/black"
-        android:textSize="24sp"
-        android:textStyle="bold"/>
-
-    <TextView
-        android:id="@+id/question1"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Color #1:"
-        android:textColor="@color/black"
-        android:textSize="20dp"
-        android:layout_marginTop="100dp" />
-
-    <RadioGroup
-        android:id="@+id/RadioGroup"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="125dp">
-        
-        <RadioButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/option1"
-            android:textSize="15dp"
-            android:text="Red"  />
-
-        <RadioButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/option2"
-            android:textSize="15dp"
-            android:text="Blue"  />
-
-        <RadioButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/option3"
-            android:textSize="15dp"
-            android:text="Yellow"  />
-    </RadioGroup>
-
-    <TextView
-        android:id="@+id/question2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="275dp"
-        android:text="Color #2:"
-        android:textColor="@color/black"
-        android:textSize="20sp" />
-
-    <Spinner
-        android:id="@+id/spinner"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="300dp"/>
-    <Button
-        android:id="@+id/button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Result"
-        android:layout_marginTop="335dp"/>
-
-    <TextView
-        android:id="@+id/result"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Based on Choices"
-        android:textColor="@color/black"
-        android:textSize="20sp"
-        android:layout_marginTop="100dp"
-        android:layout_gravity="center"/>
-
-</FrameLayout>
